requirements.txt
click >= 7.0
google-api-core == 1.16.0
google-api-python-client == 1.8.0
google-auth == 1.11.3
google-auth-httplib2 == 0.0.3
google-auth-oauthlib == 0.4.1
googleapis-common-protos == 1.51.0
gspread == 3.3.0
matplotlib == 3.2.1
numpy == 1.18.2
openpyxl == 3.0.3
pandas == 1.0.3
psycopg2-binary == 2.8.5
pyyaml == 5.3.1
reportlab == 3.5.42
sqlalchemy == 1.3.15
svglib == 1.0.0
tqdm == 4.46.0
xlrd == 1.2.0
